Remove default prod config-env from build.prod.
The names of our deployment environments (dev, test, prod) and the names
of our build targets (build.dev, build.test, and build.prod) are not in
direct correspondence. We deploy the output of build.prod to our all our
deployment environments. For example, to dev we deploy the result of `npm run
build.prod -- --config-env dev`.