GEOD-294 - Separate view model from data model
- existing behaviour: GeodesyML comes from the database
and JSONIX translates to JSON.  This is JSON with complex /
deeply nested objects where certain paths may or maynot exist
- this ticket takes the GeodesyML JSON, which we call the Data
Model, and converts it to/from a View Model as used in the UI parts
of the application.  The View Model is simpler and flatter,
avoiding the need to deeply nested objects.
- this ticket does this work for HumiditySensors and provides
a template with field partially filled out, or comments for
other fields, to allow the components for these fields to be
easily created in the same style.