GEOD-379 - Attempted fix to look at later
* Our SiteLogModel is ordered Oldest..Newest so that
diffs work properly. (data model)
* The data is displayed Newest..Oldest on the forms so
pass in reversed data (view model and form model)
* The problem comes with model driven forms as the order in the
view model is not matching the display in the form model
* I did plunkr https://plnkr.co/edit/Ftn9jHITd0hVr7S8CLin?p=preview
which highlights how to manage arrays in model driven forms
* I attempted that here but it fails.  The reason it fails is because
the array contains data model items but the corresponding
form model items are created as it loops over the data.
However with a 'formArrayName' and 'formControlName'
(see GNSS-Receiver-Group-Item.ts), it is execting the form controls
to exist.
* The solution would be to create these before-hand so they do exist
* Given this change, it might be a good time to move away from the
intermediate View Model so that we would instead have only
the Data Model and the Form Model (that would double as the View Model).
* The Group opening action would build the Form Model (or it could all
be done when the Data Model is created though this would take time
and space unnecessarily.  This could be a good way to try first.
* This separate branch has been created.  To continue any work here
would be experimental.  The original branch (reactive-model-g379)
will be changed so that the data is always Newest to Oldest.  New()
will create dummy entries in the 'orig' list so that the diff
will detect the difference properly and we'll work out the new items.
and the diff will be modified