Fix date value that wasnt appearing at ngOnInit()
* When create a new Item, it would open and the date wasn't
yet set.  So the date for the datePicker was now() rather
than the value that had been created.  So it was out-of-sync
with the date string in the input box of DateTimeComponent
* To solve this a `setTimeout()` was removed on the setValue().
It was there to protect against a 'was false now true' error.
That seems to have been fixed in other code improvements we
have made.
* Fixed test init