GEOD-407 - Validation components
* Improved Validation messages
* Fixed number custom component to take a max value and validate
against min and max

* The use of Validators.required causes a "was false now true"
error, however this seems more of a warning - we must have fixed
enough to prevent this being a fatal error.  I couldn't find a
way to prevent this 'warning'.
* Not done was Validation for DateTime since ControlValueAccessor
validation won't work for this unless it is modified.  HOWEVER
I made some improvements to the DateTime validation